requirements.txt
fastapi
uvicorn[standard]
sqlmodel
httpx
python-dotenv
python-multipart
sightengine
pydantic[email]
requests
sendgrid

